% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyRandomForest.R
\name{applyRandomForest}
\alias{applyRandomForest}
\title{Apply random forest.}
\usage{
applyRandomForest(
  dataTrain = NULL,
  dataTest = NULL,
  frmla.f = NULL,
  outcome = "y"
)
}
\arguments{
\item{dataTrain}{Training subset, e.g., 80 percent, of the full sample.}

\item{dataTest}{Test subset, e.g., 20 percent, of the full sample.}

\item{frmla.f}{An object of the class \code{formula}, used for the random forest model (f = outcome must be of class \code{factor}).}

\item{outcome}{Character that represents the column name in the data.frame which contains the observed outcome (default: 'observed').}
}
\value{
a list with two data.frames as elements (names: ApparentCV, TestCV), each data.frame having three columns:
\enumerate{
\item observed Observed outcome (0 = absent, 1 = present).
\item predicted Model-based probability estimation of the outcome being present.
\item ids Row numbers of the total sample (before data has been split into training and test subsets).
}
}
\description{
Apply the default random forest model to the data.
}
\examples{
# 2 seeds and 5-fold cross-validation would lead to 10 performance results.
set.seed(1)
seeds <- sample(1:10e6, size=2)
cvLs <- myRepeatedkFoldcv(data=dummyData, outcome="y", folds = 5, seeds=seeds)
# Extract training and test subsets of the full dataset:
TrainLs <- cvLs$TrainLs
TestLs <- cvLs$TestLs
fmla.f <- formula(factor(y) ~ .)
applyRandomForest(dataTrain = TrainLs[[1]], dataTest = TestLs[[1]], frmla=fmla.f, outcome="y")
}
\references{
\insertRef{ranger2017}{mysml}
}
\author{
Marcel Miché
}
