% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avrgCalibPlotLs.R
\name{avrgCalibPlotLs}
\alias{avrgCalibPlotLs}
\title{Make an averaged calibration plot.}
\usage{
avrgCalibPlotLs(calibPlotLs = NULL, model = "noModel")
}
\arguments{
\item{calibPlotLs}{Object that is the output from the function 'makeCalibPlotLs'.}

\item{model}{The name of the prediction model, as shown in the output of the function 'makeCalibPlotLs' or 'avrgCalibPlotLs'.}
}
\value{
a list with three data.frames as elements (names: mn, lci, uci), each data.frame having six columns:
\enumerate{
\item mn_x Mean x-axis value (x-axis = predicted probabilities).
\item mn_y  Mean y-axis value (y-axis = observed outcome rate).
\item lci_y Lower limit of the 95 percent confidence interval.
\item uci_y Upper limit of the 95 percent confidence interval.
\item model Name of the prediction model.
\item tile Enumeration of selected number of approximately equal sized subsamples.
}
}
\description{
Average results across several calibration plots.
}
\examples{
# Use the saved output from the function 'computeRelevantResults'.
# Before 'avrgCalibPlotLs' can be used, use 'makeCalibPlotLs':
calibLRLs <- makeCalibPlotLs(calibLs=test$orderedObsLs, model="logreg")
# Use this output to continue with 'avrgCalibPlotLs':
testcalibLR <- avrgCalibPlotLs(calibPlotLs = calibLRLs, model="logreg")
# Peek at first 2 lines of output:
lapply(testcalibLR, function(x) head(x, n=2))
# $mn
#           mn_x mn_y lci_y uci_y  model tile
# 1 0.0005528451    0     0     0 logreg    1
# 2 0.0015136628    0     0     0 logreg    2
# 
# $lci
#           mn_x mn_y lci_y uci_y  model tile
# 1 0.0004820103    0     0     0 logreg    1
# 2 0.0013561840    0     0     0 logreg    2
# 
# $uci
#          mn_x mn_y lci_y uci_y  model tile
# 1 0.000623680    0     0     0 logreg    1
# 2 0.001671142    0     0     0 logreg    2
}
\author{
Marcel Miché
}
