% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRelevantResults.R
\name{computeRelevantResults}
\alias{computeRelevantResults}
\title{Relevant results for a clinical prediction model.}
\usage{
computeRelevantResults(
  predictionOutputLs = NULL,
  dcaStartAtZero = TRUE,
  dcaReasonableThresholds = c(0.01, 0.02, 0.03, 0.04, 0.05),
  fullModelNames = c("Logistic regression")
)
}
\arguments{
\item{predictionOutputLs}{A list whose elements are data.frames, each data.frame containing a column with observed events (0 = no, 1 = yes) and a column with the predicted probabilities of observing the event.}

\item{dcaStartAtZero}{Boolean value. TRUE (default) if the x-axis of the decision curve plot shall include zero, else FALSE.}

\item{dcaReasonableThresholds}{Numeric vector with the selected reasonable range of threshold probabilities.}

\item{fullModelNames}{Character vector with the full names of the prediction model, e.g., Logistic regression instead of logreg. There must be as many full model names as the length of the argumet \code{predictionOutputLs}, and they must be in the same order.}
}
\value{
a list with three lists as elements:
\enumerate{
\item orderedObsLs A list that contains data.frames (df), each df contains the column observed (= observed outcome, 0 = no, 1 = yes) and the column predicted (= predicted probabilities).
\item dcaLs Two lists, named tableDCA and plotDCA, respectively (see \strong{Examples}).
\item calibLs A list that contains data.frames (df), each df contains the predicted probabilities of all prediction models (column name is the prediction model, e.g., logreg) and the observed outcome (column name 'observed').
}
}
\description{
Compute results that are relevant for a clinical prediction model.
}
\examples{
# Use the predicted probabilities (output for several cross-validations
# of two prediction models, logistic regression and random forest), which
# are computed from using a simulated dummy dataset.
test <- computeRelevantResults(predictionOutputLs = predProbsLs,
                               dcaReasonableThresholds = c(.02, .03, .04, .05),
            fullModelNames = c("Logistic regression",
                              "Random forest"))
}
\references{
\insertRef{vickers2006decision}{mysml}

\insertRef{vickers2019simple}{mysml}

\insertRef{van2018reporting}{mysml}

\insertRef{van2016calibration}{mysml}
}
\author{
Marcel Miché
}
