% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_plotfun.R
\name{dca_plotfun}
\alias{dca_plotfun}
\title{Decision curve analysis functionality.}
\usage{
dca_plotfun(y = NULL, p = NULL, dcaReasonableThresholds = NULL)
}
\arguments{
\item{y}{Vector containing integer values 0 and 1 (0 = event absent, 1 = event present).}

\item{p}{Vector containing probabilities, i.e., continuous values between 0 and 1.}

\item{dcaReasonableThresholds}{Numeric vector with the selected reasonable range of threshold probabilities.}
}
\value{
a data.frame with three columns:
\enumerate{
\item label Prediction model.
\item threshold The selected reasonable threshold probabilities.
\item net_benefit The net benefit for the respective threshold.
}
}
\description{
Compute decision curve analysis output.
}
\examples{
dca_plotfun(y=rep(c(0,1), times=c(10, 10)),
p=c(.03, .06, .09, .11, .35, .05, .3, .02, .15, .11,
    .2, .39, .13, .4, .44, .07, .22, .31, .32, .5),
dcaReasonableThresholds=c(.15, .2, .25))
}
\references{
\insertRef{vickers2006decision}{mysml}

\insertRef{vickers2019simple}{mysml}

\insertRef{van2018reporting}{mysml}

\insertRef{van2016calibration}{mysml}
}
\author{
Marcel Miché
}
