% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htbPlotVec.R
\name{htbPlotVec}
\alias{htbPlotVec}
\title{Harm component of harm-to-benefit.}
\usage{
htbPlotVec(x)
}
\arguments{
\item{x}{Numeric values, representing probabilities, here excluding the values 0 and 1.}
}
\value{
A numeric value that represents the harm component of the harm-to-benefit ratio.
}
\description{
Compute the 'harm' component of the 'harm-to-benefit' ratio.
}
\examples{
# Harm component of harm-to-benefit ratio for a probability of .25 is
# 3: .25/(1-.25) = .25:.75 = 1:3 (harm-to-benefit ratio = 1:3).
htbPlotVec(x=c(.05, .2, .25)) # Output: 19, 4, 3
}
\author{
Marcel Miché
}
