% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htbToProb.R
\name{htbToProb}
\alias{htbToProb}
\title{Probabilities from given harm-to-benefit ratios.}
\usage{
htbToProb(x = "1:9")
}
\arguments{
\item{x}{Character that represents a harm-to-benefit ratio, such as (default) '1:9'.}
}
\value{
A numeric value that represents the probability that corresponds to the given harm-to-benefit ratio.
}
\description{
Compute the probability for a given harm-to-benefit ratio.
}
\examples{
# Reverse a harm-to-benefit ratio to the corresponding probability:
htbToProb(x=c("1:3", "1:9")) # Output: 0.25, 0.10
}
\author{
Marcel Miché
}
