% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCalibPlotLs.R
\name{makeCalibPlotLs}
\alias{makeCalibPlotLs}
\title{Results to produce a calibration plot.}
\usage{
makeCalibPlotLs(calibLs = NULL, model = "noModel", n = 10)
}
\arguments{
\item{calibLs}{Object that is the first out of three list elements (output) from function 'computeRelevantResults'.}

\item{model}{The name of the prediction model, as shown in the output of the function 'makeCalibPlotLs' or 'avrgCalibPlotLs'.}

\item{n}{Integer value that determines the number of roughly equal sized groups of the test data subset (default n = 10).}
}
\value{
a data.frame with seven columns:
\enumerate{
\item mn_x Mean x-axis value (x-axis = predicted probabilities).
\item mn_y  Mean y-axis value (y-axis = observed outcome rate).
\item lci_y Lower limit of the 95 percent confidence interval.
\item uci_y Upper limit of the 95 percent confidence interval.
\item model Name of the prediction model.
\item run The run among all of the cross-validations of the prediction model.
\item tile Enumeration of selected number of approximately equal sized subsamples.
}
}
\description{
Compute results that are necessary for a single calibration plot.
}
\examples{
# Use part of the saved output from the function 'computeRelevantResults'.
calibLRLs <- makeCalibPlotLs(calibLs=test$orderedObsLs, model="logreg")
# Peek at first 2 lines of output:
head(calibLRLs[[1]], n = 2)
#          mn_x mn_y lci_y uci_y  model run tile
# 1 0.000508724    0     0     0 logreg   1    1
# 2 0.001480403    0     0     0 logreg   1    2
}
\references{
\insertRef{van2016calibration}{mysml}
}
\author{
Marcel Miché
}
