% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnci.R
\name{mnci}
\alias{mnci}
\title{Mean and percantage confidence interval.}
\usage{
mnci(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric values.}

\item{alpha}{Numeric value that represents the selected alpha statistical significance level, default is 0.05.}
}
\value{
a named vector with three elements:
\enumerate{
\item mn Mean.
\item lci lower limit of the 95 percent confidence interval.
\item uci upper limit of the 95 percent confidence interval.
}
}
\description{
Compute mean with some percentage confidence interval.
}
\examples{
mnci(x=50:60)
}
\references{
Recommended link \href{https://matheusfacure.github.io/python-causality-handbook/03-Stats-Review-The-Most-Dangerous-Equation.html}{The most dangerous equation}.
}
\author{
Marcel Miché
}
