% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnminmax.R
\name{mnminmax}
\alias{mnminmax}
\title{Mean, minimum, and maximum.}
\usage{
mnminmax(x)
}
\arguments{
\item{x}{Numeric values.}
}
\value{
a named vector with three elements:
\enumerate{
\item mn Mean.
\item min Minimum.
\item max Maximum.
}
}
\description{
Compute mean with a minimum to maximum interval.
}
\examples{
mnminmax(x=50:60)
}
\author{
Marcel Miché
}
