% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRepeatedkFoldcv.R
\name{myRepeatedkFoldcv}
\alias{myRepeatedkFoldcv}
\title{Repeated k-fold cross-validation.}
\usage{
myRepeatedkFoldcv(
  data = NULL,
  outcome = "y",
  folds = NULL,
  stratify = TRUE,
  seeds = NULL
)
}
\arguments{
\item{data}{The study data that is used for the prediction modeling.}

\item{outcome}{Character that represents the column name in the data.frame which contains the observed outcome (default: 'observed').}

\item{folds}{Numeric, the number of folds in the repeated k-fold cross-validation.}

\item{stratify}{Boolean value to determine whether the outcome rate in the cross-validation subsamples shall be (roughly) the same as in the total sample (TRUE) or whether the outcome rate may vary (FALSE).}

\item{seeds}{Numeric vector with a least one seed value (see \strong{Details}).}
}
\value{
a list with two lists as elements:
\enumerate{
\item TrainLs A list with all training subsets of the data.
\item TestLs A list with all test subsets of the data.
}
}
\description{
Setup a repeated k-fold cross-validation procedure.
}
\details{
A seed is a numeric value with which the user can guarantee that a sampling function yields reproducible output, independent of who runs the sampling function or when it is used.
}
\examples{
# 5 seeds and 5-fold cross-validation leads to 25 performance results.
set.seed(1)
seeds <- sample(1:10e6, size=5)
cvLs <- myRepeatedkFoldcv(data=dummyData, outcome="y", folds = 5, seeds=seeds)
# Extract training and test subsets of the full dataset:
TrainLs <- cvLs$TrainLs
TestLs <- cvLs$TestLs
}
\author{
Marcel Miché
}
