% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysml.R
\docType{package}
\name{mysml-package}
\alias{mysml-package}
\title{Documentation of this mysml package.}
\description{
This is an R package that the package author currently considers useful for applying supervised machine learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmiche/mysml}
  \item Report bugs at \url{https://github.com/mmiche/mysml/issues}
}

}
