% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsToProb.R
\name{oddsToProb}
\alias{oddsToProb}
\title{Transform odds to probabilities.}
\usage{
oddsToProb(x)
}
\arguments{
\item{x}{Numeric value that represents odds, e.g., 0.25 (resuling from 0.2/(1-0.2)).}
}
\value{
The probability that is computed from the given odds.
}
\description{
Compute probabilities from odds.
}
\examples{
oddsToProb(x=c(.1, .25)) # Output: 0.09090909 0.20000000
}
\references{
\insertRef{greenland1998probability}{mysml}
}
\author{
Marcel Miché
}
