% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probToOdds.R
\name{probToOdds}
\alias{probToOdds}
\title{Transform probabilities to odds.}
\usage{
probToOdds(x)
}
\arguments{
\item{x}{Numeric value that represents a probability, i.e., a number between 0 and 1.}
}
\value{
The odds that is computed from the given probability.
}
\description{
Compute odds from probabilities.
}
\examples{
probToOdds(x=c(.1, .2)) # Output: 0.1111111 0.2500000
}
\references{
\insertRef{greenland1998probability}{mysml}
}
\author{
Marcel Miché
}
