% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translatenb.R
\name{translatenb}
\alias{translatenb}
\title{Translate net benefit.}
\usage{
translatenb(nb = NULL)
}
\arguments{
\item{nb}{Numeric values that represent a positive net benefit result, i.e., values between > 0 and < 1 (see \strong{Details}).}
}
\value{
A list with a named vector for each given net benefit (nb), i.e., its translation. For example, nb = 0.1 returns a count of 1 (true positive) per 10 (predicted probabilities). In words: A net benefit of 0.1 means that if the prediction model is used as the clinical decision strategy, then one true positive (TP) individual is among every 10 individuals whose outcome risk has been predicted with the clinical prediction model. This one TP individual is obtained without any further increase of false positive (FP) individuals (which is the reason why it is called 'net' benefit).
}
\description{
Translate a net benefit result.
}
\details{
The theoretical maximum net benefit is 1, which is irrelevant for this translation function.
}
\examples{
translatenb(nb=c(.003, .15))
# Output:
# [[1]]
#      nb count   per
# 1 0.003     3 1'000
# 
# [[2]]
#     nb count per
# 1 0.15    15 100
}
\author{
Marcel Miché
}
