% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMultipleDCA.R
\name{visualizeMultipleDCA}
\alias{visualizeMultipleDCA}
\title{Prepare data for plotDCA.}
\usage{
visualizeMultipleDCA(dcaLs = NULL, dcaSelectType = NULL)
}
\arguments{
\item{dcaLs}{The output dcaLs from function \code{computeRelevantResults} from the \code{mysml} package.}

\item{dcaSelectType}{Character string. One out of these three options (see \strong{Details}): mnci, mnminmax, mnq1q3.}
}
\value{
a list with three elements:
\enumerate{
\item allDCA A dataframe needed by the function \code{plotDCA}.
\item logregNB A dataframe with as many rows as selected threshold probabilities. The columns contain summaries of the net benefit, i.e., the median, the mean, the minimum, the first and the third quantile, and the maximum.
\item rfNB Same as logregNB, only for the random forest model.
}
}
\description{
Function that prepares the data for the function plotDCA.
}
\examples{
# Assuming that the output of mysml::computeRelevantResults has
# been assigned to the variable 'relRes'. Then continue with:
# visualizeMultipleDCA(dcaLs = relResLR$dcaLs, dcaSelectType = "mnci")
}
\references{
\insertRef{dplyr2023}{mysml}

\insertRef{magrittr2022}{mysml}
}
\author{
Marcel Miché
}
