% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_stability1.R
\name{calibration_stability1}
\alias{calibration_stability1}
\title{Prepare calibration stability plot data.}
\usage{
calibration_stability1(x)
}
\arguments{
\item{x}{Output of function \code{validate} from the \code{pminternal} package.}
}
\value{
a list with b + 1 elements (b = number of bootstrap repetitions). Each element is a dataframe with 100 rows and two columns:
\enumerate{
\item p predicted probabilities.
\item pc calibrated probabilities.
}
The 100 rows of each dataframe stem from the default number of evaluations from the function \code{pmcalibration::pmcalibration}, i.e., generating 100 equally spaced intervals between the minimum and the maximum estimated probability. Furthermore, the default smoothing function \code{mgcv::gam} (generalized additive model) is used to calibrate the probabilities \code{p}, which produces the column \code{pc}.
See also Riley and Collins, 2023, section 3.2.2.
}
\description{
Extract data for calibration stability.
}
\details{
This function is a modification of the function \code{calibration_stability} of the R package \code{pminternal}. The reason to modify it, was that I wanted to visualize the results with ggplot2, not with the R graphics package.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Calibration stability.
}
\references{
\insertRef{pmcalibration2025}{mysml}

\insertRef{pminternal2025}{mysml}

\insertRef{riley2023stability}{mysml}

\insertRef{collins2024evaluation}{mysml}
}
\author{
Stephen Rhodes (see \strong{Details})
}
