% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelevantBootstrapOutputLs.R
\name{getRelevantBootstrapOutputLs}
\alias{getRelevantBootstrapOutputLs}
\title{Get relevant output from the bootstrap cross-validation procedure.}
\usage{
getRelevantBootstrapOutputLs(
  x = NULL,
  listName = "modelName",
  withApparent = FALSE
)
}
\arguments{
\item{x}{Output of function \code{validate} from the \code{pminternal} package.}

\item{listName}{Character string. Name of the model (default: "modelName").}

\item{withApparent}{Boolean value (default: FALSE). If apparent predicted probabilities shall be included, change argument to TRUE.}
}
\value{
a list with b + 1 or b + 2 elements (b = number of bootstrap repetitions). Each element is a dataframe with as many rows as in the original sample size and two columns:
\enumerate{
\item observed Binary numeric values 0 (no event observed) and 1 (event observed).
\item predicted Numeric value between 0 and 1 (estimated probability that event was observed).
}
The apparent predicted probabilities (from the original sample) can either be included or (this is the default) excluded.
}
\description{
Extract data for performance visualization.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Decision curve stability.
}
\references{
\insertRef{pminternal2025}{mysml}
}
\author{
Stephen Rhodes (see \strong{Details})
}
