% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCalib.R
\name{myCalib}
\alias{myCalib}
\title{Calibration related summary results.}
\usage{
myCalib(calibDf = NULL, outcome = "observed")
}
\arguments{
\item{calibDf}{Object (a data.frame) that is the second out of three list elements (output) from function 'computeRelevantResults'.}

\item{outcome}{Character that represents the column name in the data.frame which contains the observed outcome (default: 'observed').}
}
\value{
a list with three data.frames (df) as elements:
\enumerate{
\item calibDf Return the same data that was received by the function.
\item Return calibration related results (see \strong{Details}) in long format.
\item Return calibration related results (see \strong{Details}) in wide format.
}
}
\description{
Compute some calibration related summary results.
}
\details{
The returned calibration related results are: 
\enumerate{
\item Brier score: Best possible result (Bpr) = 0.
\item Calibration Intercept: Bpr = 0.
\item Calibration Slope: Bpr = 1.
\item Emax: Bpr = 0.
\item E90: Bpr = 0.
\item ECE: Bpr = 0.
\item MCE: Bpr = 0.
\item ICI: Bpr = 0.
}
}
\examples{
# Use part of the saved output from the function 'computeRelevantResults'.
calibResultsLs <- lapply(test$calibLs, FUN=function(x) {
    myCalib(calibDf = x, outcome="observed")
})
# Peek at part of output:
calibResultsLs[[1]]$calibPerfW[,c("Intercept", "Slope", "ici", "names")]
#              Intercept     Slope         ici        names
# logreg        0.142350 0.9843143 0.005361289       logreg
# randomForest  1.107592 1.3395060 0.006857659 randomForest
}
\author{
Marcel Miché
}
