% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysml.R
\docType{package}
\name{mysml}
\alias{mysml}
\title{Documentation of this mysml package.}
\description{
This is an R package that the package author currently considers useful for applying supervised machine learning.

All R packages that have been used in this mysml package are listed below (see References).

Finally, R packages that were used to develop and publish this mysml R package on GitHub are: devtools, usethis, roxygen2, and Rdpack. Thanks to the many R package developers!
}
\references{
\insertRef{tidyverse}{mysml}

\insertRef{tibble}{mysml}

\insertRef{magrittr}{mysml}

\insertRef{devtools}{mysml}

\insertRef{usethis}{mysml}

\insertRef{roxygen2}{mysml}

\insertRef{Rdpack}{mysml}
}
