% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_stability1.R
\name{prediction_stability1}
\alias{prediction_stability1}
\title{Prepare prediction stability plot data.}
\usage{
prediction_stability1(x, bounds = 0.95, smooth_bounds = FALSE, span = 0.75)
}
\arguments{
\item{x}{Output of function \code{validate} from the \code{pminternal} package.}

\item{bounds}{Width of the 'stability interval' (percentiles of the bootstrap model predictions). NULL = do not add bounds to plot.}

\item{smooth_bounds}{Boolean value (default = FALSE). If TRUE, uses loess to smooth the bounds; see Argument \code{smooth_bounds} in \code{pminternal::prediction_stability}.}

\item{span}{Numeric value. Controls the degree of smoothing (see loess; default = 0.75); see Argument \code{span} in \code{pminternal::prediction_stability}.}
}
\value{
a list with three lists as elements:
\enumerate{
\item b Number of bootstrap repetitions (we used b = 500).
\item stabilDf Dataframe with as many rows as the original data and with b + 1 columns (column 1 = p_app = apparent predicted probability, whereas the remaining b many columns contain the cross-validated predicted probabilities).
\item lims Dataframe with as many rows as the original data and with three columns (names: X2.5., X.97.5, and p_app). In the prediction instability plot, they are shown as the dashed 95 percent stability interval for the estimated risks from the bootstrap models (see Riley and Collins, 2023, section 3.2.1).
}
}
\description{
Extract data for prediction stability.
}
\details{
This function is a modification of the function \code{prediction_stability} of the R package \code{pminternal}. The reason to modify it, was that I wanted to visualize the results with ggplot2, not with the R graphics package.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Prediction stability.
}
\references{
\insertRef{pminternal2025}{mysml}
}
\author{
Stephen Rhodes (see \strong{Details})
}
